Renamed day15_part1_runner to day15_runner
