+++ Makefile
@@ -220,17 +220,23 @@ build/programs/day15_area.o: src/programs/day15_area.cpp  \
 	include/programs/day15_point.h
 	g++ ${CPPFLAGS} -o build/programs/day15_area.o -c src/programs/day15_area.cpp
 
-build/programs/day15_part1_runner.o: src/programs/day15_part1_runner.cpp  \
-	include/programs/day15_part1_runner.h \
+build/programs/day15_runner.o: src/programs/day15_runner.cpp  \
+	include/programs/day15_runner.h \
 	include/programs/day15_area.h \
 	include/programs/day15_point.h
-	g++ ${CPPFLAGS} -o build/programs/day15_part1_runner.o -c src/programs/day15_part1_runner.cpp
+	g++ ${CPPFLAGS} -o build/programs/day15_runner.o -c src/programs/day15_runner.cpp
 
 build/programs/day15_part1.o: src/programs/day15_part1.cpp  \
-	include/programs/day15_part1_runner.h \
+	include/programs/day15_runner.h \
 	include/programs/day15_area.h \
 	include/programs/day15_point.h
 	g++ ${CPPFLAGS} -o build/programs/day15_part1.o -c src/programs/day15_part1.cpp
+
+build/programs/day15_part2.o: src/programs/day15_part2.cpp  \
+	include/programs/day15_runner.h \
+	include/programs/day15_area.h \
+	include/programs/day15_point.h
+	g++ ${CPPFLAGS} -o build/programs/day15_part2.o -c src/programs/day15_part2.cpp
 
 build/test/test_memoryloader.o: src/test/test_memoryloader.cpp  \
 	include/core/memory_loader.h  \
@@ -397,11 +403,20 @@ bin/programs/day13_part2: build/programs/day13_part2.o  \
 bin/programs/day15_part1: build/programs/day15_part1.o  \
 	build/programs/day15_point.o  \
 	build/programs/day15_area.o  \
-	build/programs/day15_part1_runner.o  \
+	build/programs/day15_runner.o  \
 	bin/lib/libcore.a \
 	bin/lib/liboperations.a \
 	bin/lib/libprograms.a
-	g++ ${CPPFLAGS} -o bin/programs/day15_part1 build/programs/day15_part1.o build/programs/day15_point.o build/programs/day15_area.o build/programs/day15_part1_runner.o -Lbin/lib -lprograms -loperations -lcore
+	g++ ${CPPFLAGS} -o bin/programs/day15_part1 build/programs/day15_part1.o build/programs/day15_point.o build/programs/day15_area.o build/programs/day15_runner.o -Lbin/lib -lprograms -loperations -lcore
+
+bin/programs/day15_part2: build/programs/day15_part2.o  \
+	build/programs/day15_point.o  \
+	build/programs/day15_area.o  \
+	build/programs/day15_runner.o  \
+	bin/lib/libcore.a \
+	bin/lib/liboperations.a \
+	bin/lib/libprograms.a
+	g++ ${CPPFLAGS} -o bin/programs/day15_part2 build/programs/day15_part2.o build/programs/day15_point.o build/programs/day15_area.o build/programs/day15_runner.o -Lbin/lib -lprograms -loperations -lcore
 
 bin/test/test_memoryloader: build/test/test_memoryloader.o  \
 	build/core/memory_loader.o  \
@@ -496,8 +511,9 @@ clean:
 	build/programs/day13_part2.o  \
 	build/programs/day15_point.o  \
 	build/programs/day15_area.o  \
-	build/programs/day15_part1_runner.o  \
+	build/programs/day15_runner.o  \
 	build/programs/day15_part1.o  \
+	build/programs/day15_part2.o  \
 	build/test/test_memoryloader.o  \
 	build/test/test_operation_addition.o  \
 	build/test/test_operation_multiplication.o  \
@@ -522,6 +538,7 @@ clean:
 	bin/programs/day13_part1 \
 	bin/programs/day13_part2 \
 	bin/programs/day15_part1 \
+	bin/programs/day15_part2 \
 	bin/test/test_memoryloader \
 	bin/test/test_operation_addition \
 	bin/test/test_operation_multiplication \
@@ -572,8 +589,9 @@ all: build/core/memory_page.o  \
 	build/programs/day13_part2.o  \
 	build/programs/day15_point.o  \
 	build/programs/day15_area.o  \
-	build/programs/day15_part1_runner.o  \
+	build/programs/day15_runner.o  \
 	build/programs/day15_part1.o  \
+	build/programs/day15_part2.o  \
 	build/test/test_memoryloader.o  \
 	build/test/test_operation_addition.o  \
 	build/test/test_operation_multiplication.o  \
@@ -596,6 +614,7 @@ all: build/core/memory_page.o  \
 	bin/programs/day13_part1 \
 	bin/programs/day13_part2 \
 	bin/programs/day15_part1 \
+	bin/programs/day15_part2 \
 	bin/test/test_operation_addition \
 	bin/test/test_operation_multiplication \
 	bin/test/test_operation_input_output \
